address copilot review feedback
1. Move rxjs dependency from root package.json to apps/notebook/package.json
   where the actual imports live. Fixes workspace resolution.

2. Extract CellChangeset types + mergeChangesets to lib/cell-changeset.ts
   (pure module, zero external deps). Tests now import from the pure module
   without pulling in Tauri/RxJS/WASM runtime. frame-pipeline.ts re-exports
   for downstream consumers.

3. Fix overlapping async materializations: replace subscribe-with-async
   callback with concatMap in the coalescing buffer pipeline. Ensures
   batches that take longer than COALESCE_MS (e.g. blob resolution)
   queue and execute serially rather than racing store writes.

4. Fix initial sync reply ordering: wrap initial materializeCells in
   concatMap so onSyncApplied() is called AFTER the store is populated.
   Prevents sending a sync reply while the cell store is still empty,
   which could cause incremental updateCellById calls to be dropped.
