+++ package.json
@@ -69,7 +69,6 @@
     "rehype-raw": "^7.0.0",
     "remark-gfm": "^4.0.1",
     "remark-math": "^6.0.0",
-    "rxjs": "^7.8.2",
     "tailwind-merge": "^3.4.0"
   },
   "devDependencies": {
